Bug 1928619 - test262-export: don't export the main shell.js and move the shell.js files to a subdirectory; r=dminor
Differential Revision: https://phabricator.services.mozilla.com/D227651

+++ js/src/tests/test262-export.py
@@ -34,6 +34,7 @@ SUPPORT_FILES = set(
 def findAndCopyIncludes(dirPath: str, baseDir: str, includeDir: str) -> "list[str]":
     relPath = os.path.relpath(dirPath, baseDir)
     includes: list[str] = []
+    os.makedirs(os.path.join(includeDir, "sm"), exist_ok=True)
 
     # Recurse down all folders in the relative path until
     # we reach the base directory of shell.js include files.
@@ -42,28 +43,16 @@ def findAndCopyIncludes(dirPath: str, baseDir: str, includeDir: str) -> "list[st
         # find the shell.js
         shellFile = os.path.join(baseDir, relPath, "shell.js")
 
-        # create new shell.js file name
-        includeFileName = relPath.replace("/", "-") + "-shell.js"
-        includesPath = os.path.join(includeDir, includeFileName)
-
-        if os.path.exists(shellFile):
-            # if the file exists, include in includes
+        # if the file exists and is not empty, include in includes
+        if os.path.exists(shellFile) and os.path.getsize(shellFile) > 0:
+            # create new shell.js file name
+            includeFileName = "sm/" + relPath.replace("/", "-") + "-shell.js"
             includes.append(includeFileName)
 
-            if not os.path.exists(includesPath):
-                shutil.copyfile(shellFile, includesPath)
+            includesPath = os.path.join(includeDir, includeFileName)
+            shutil.copyfile(shellFile, includesPath)
 
         relPath = os.path.split(relPath)[0]
-
-    shellFile = os.path.join(baseDir, "shell.js")
-    includesPath = os.path.join(includeDir, "shell.js")
-    if not os.path.exists(includesPath):
-        shutil.copyfile(shellFile, includesPath)
-
-    includes.append("shell.js")
-
-    if not os.path.exists(includesPath):
-        shutil.copyfile(shellFile, includesPath)
 
     return includes
